WIP: Use AVFoundation without using AVPlayer
This is needed to enable Multimedia player support on iOS.  Instead of
using the highlevel AVPlayer API, and render multimedia to AVPlayerLayer
objects, we use AVAssetReader and decode and play the media ourselves.
This allows us to have full control over both the decoded audio and
video streams.

This is only a commit of the work I've done so far.  The Video should
play without audio, and only from local URLs.

Change-Id: I8515017464a4daf3c0423bc9278c424a92cc53af